new object vs data analogy
First off, 'legos' is not a game.  And nobody takes apart their legos just because it's bedtime.  Also this analogy breaks down because you say the important thing is that we're not storing our object in the database, but you *are* putting the legos in the box.

I've attempted a new analogy.  Take it or leave it.